+++ src/main/java/org/cotato/csquiz/common/config/WebConfig.java
@@ -1,8 +1,8 @@
 package org.cotato.csquiz.common.config;
 
 import lombok.RequiredArgsConstructor;
-import org.cotato.csquiz.common.interceptor.IdempotencyInterceptor;
-import org.cotato.csquiz.common.interceptor.IdempotencyRedisRepository;
+import org.cotato.csquiz.common.idempotency.IdempotencyInterceptor;
+import org.cotato.csquiz.common.idempotency.IdempotencyRedisRepository;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
