+++ src/main/java/org/cotato/csquiz/common/idempotency/IdempotencyResponse.java
@@ -1,4 +1,4 @@
-package org.cotato.csquiz.common.interceptor;
+package org.cotato.csquiz.common.idempotency;
 
 import java.io.Serializable;
 import lombok.AllArgsConstructor;
